+++ Samples/Sample.Console/Program.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
+using ListBox = Platform.Support.ConsoleEx.ListBox;
 
 namespace Sample.Console
 {
@@ -22,6 +23,13 @@ namespace Sample.Console
             if (Platform.Support.OS.OSHelper.IsLinux())
                 System.Console.WriteLine("Running Linux");
 
+            System.Console.ReadKey();
+
+            string[] months = { "January", "February", "March", "April", "May", "June", "July", "August", "September", "October", "November", "December" };
+            ListBox.WriteColorString("Choose Level using down and up arrow keys and press enter", 12, 20, ConsoleColor.Black, ConsoleColor.White);
+            int choice = ListBox.ChooseListBoxItem(months, 34, 3, ConsoleColor.Blue, ConsoleColor.White);
+            // do something with choice
+            ListBox.WriteColorString("You chose " + months[choice - 1] + ". Press any key to exit", 21, 22, ConsoleColor.Black, ConsoleColor.White);
 
             System.Console.ReadKey();